+++ assets/js/dashboard/stats/modals/modal.js
@@ -3,12 +3,8 @@ import { createPortal } from 'react-dom'
 import { isModifierPressed, isTyping, Keybind } from '../../keybinding'
 import { rootRoute } from '../../router'
 import { useAppNavigate } from '../../navigation/use-app-navigate'
-
-// This corresponds to the 'md' breakpoint on TailwindCSS.
-const MD_WIDTH = 768
 // We assume that the dashboard is by default opened on a desktop. This is also a fall-back for when, for any reason, the width is not ascertained.
 const DEFAULT_WIDTH = 1080
-
 class Modal extends React.Component {
   constructor(props) {
     super(props)
@@ -27,26 +23,21 @@ class Modal extends React.Component {
     window.addEventListener('resize', this.handleResize, false)
     this.handleResize()
   }
-
   componentWillUnmount() {
     document.body.style.overflow = null
     document.body.style.height = null
     document.removeEventListener('mousedown', this.handleClickOutside)
     window.removeEventListener('resize', this.handleResize, false)
   }
-
   handleClickOutside(e) {
     if (this.node.current.contains(e.target)) {
       return
     }
-
     this.props.onClose()
   }
-
   handleResize() {
     this.setState({ viewport: window.innerWidth })
   }
-
   /**
    * @description
    * Decide whether to set max-width, and if so, to what.
@@ -56,12 +47,11 @@ class Modal extends React.Component {
    */
   getStyle() {
     const { maxWidth } = this.props
-    const { viewport } = this.state
     const styleObject = {}
     if (maxWidth) {
       styleObject.maxWidth = maxWidth
     } else {
-      styleObject.width = viewport <= MD_WIDTH ? 'min-content' : '860px'
+      styleObject.maxWidth = '880px'
     }
     return styleObject
   }
@@ -78,16 +68,17 @@ class Modal extends React.Component {
         />
         <div className="modal is-open" onClick={this.props.onClick}>
           <div className="modal__overlay">
-            <button className="modal__close"></button>
-            <div
-              ref={this.node}
-              className="modal__container dark:bg-gray-900 focus:outline-hidden"
-              style={this.getStyle()}
-              // eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex
-              tabIndex={0}
-            >
-              <FocusOnMount focusableRef={this.node} />
-              {this.props.children}
+            <div className="[--gap:1rem] sm:[--gap:2rem] md:[--gap:4rem] flex h-full w-full items-center md:items-start justify-center p-[var(--gap)] box-border">
+              <div
+                ref={this.node}
+                className="max-h-[calc(100dvh_-_var(--gap)*2)] min-h-[66vh] md:min-h-120 w-full flex flex-col bg-white p-3 md:px-6 md:py-4 overflow-hidden box-border transition-[height] duration-200 ease-in shadow-2xl rounded-lg dark:bg-gray-900 focus:outline-hidden"
+                style={this.getStyle()}
+                // eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex
+                tabIndex={0}
+              >
+                <FocusOnMount focusableRef={this.node} />
+                {this.props.children}
+              </div>
             </div>
           </div>
         </div>